+++ src/main/java/com/yeoljeong/tripmate/record/presentaion/controller/RecordController.java
@@ -1,17 +1,22 @@
-package com.yeoljeong.tripmate.record.presentaion;
+package com.yeoljeong.tripmate.record.presentaion.controller;
 
 import com.yeoljeong.tripmate.auth.annotation.LoginUser;
 import com.yeoljeong.tripmate.auth.context.UserContext;
+import com.yeoljeong.tripmate.record.application.service.command.RecordCommandService;
 import com.yeoljeong.tripmate.record.presentaion.dto.request.FeedCreateRequest;
-import com.yeoljeong.tripmate.record.presentaion.dto.response.FeedCreateResponse;
-import com.yeoljeong.tripmate.record.service.command.RecordCommandService;
+import com.yeoljeong.tripmate.record.presentaion.dto.request.FeedVisibilityRequest;
+import com.yeoljeong.tripmate.record.presentaion.dto.response.FeedBaseResponse;
+import com.yeoljeong.tripmate.record.presentaion.dto.response.FeedDetailResponse;
 import com.yeoljeong.tripmate.response.ApiResponse;
 import com.yeoljeong.tripmate.response.constants.CommonSuccessCode;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.PatchMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -23,16 +28,32 @@ public class RecordController {
   private final RecordCommandService recordCommandService;
 
   @PostMapping()
-  public ApiResponse<FeedCreateResponse> createFeedData(
+  public ApiResponse<FeedDetailResponse> createFeedData(
       @LoginUser UserContext userContext,
       @Validated @ModelAttribute FeedCreateRequest request
   ) {
     return ApiResponse.success(
         CommonSuccessCode.OK,
-        FeedCreateResponse.from(
+        FeedDetailResponse.from(
             recordCommandService.createFeedImage(
                 request.toCommand(UUID.fromString(userContext.userId()))
             ))
+    );
+  }
+
+  @PatchMapping("/{feedId}/visibility")
+  public ApiResponse<FeedBaseResponse> updateFeedVisibility(
+      @LoginUser UserContext userContext,
+      @PathVariable UUID feedId,
+      @Validated @RequestBody FeedVisibilityRequest request
+  ) {
+    return ApiResponse.success(
+        CommonSuccessCode.OK,
+        FeedBaseResponse.from(
+            recordCommandService.updateFeedVisibility(
+                request.toCommand(UUID.fromString(userContext.userId()), feedId)
+            )
+        )
     );
   }
 }
